requirements.txt
yfinance
requests
rich
google-genai
